attempt using float instead

+++ features/steps/walking.py
@@ -35,7 +35,7 @@ def step_impl(context):
         #f.write(log['message'][log['message'].find("x:"):log['message'].find("x:")+6][-3:-1])
     #parses the console for a the message executed above and checkes the the x value has changed
     log = context.browser.get_log("browser")
-    assert(int(log[-1]['message'][log[-1]['message'].find("x:"):log[-1]['message'].find("x:")+6][-3:-1]) > 5)
+    assert(float(log[-1]['message'][log[-1]['message'].find("x:"):log[-1]['message'].find("x:")+6][-3:-1]) > 5)
